Manual and partial transplant of a06f785963a8 :
[lso] Initial patch to add TX LSO, VXLAN and NVGRE offload. wip://NFDH-96
No support for MTU sizes over 4096 when LSO is enabled.
The compile switch to include the code for the NIC app is LSO_ENABLE=1.
There is no support for LSO specific counters to be enabled in this patch.
If the NIC app is built with LSO enabled you must load the Linux driver that
supports LSO also.
The feature has been tested against the non-LSO driver successfully with LSO
unadvertised.
There will be more optimizations and code clean up to come in more patches.