[svc] clear pending interrupts when queue is disabled
I observed while running ifup; ifdown in a loop under heavy traffic
that interrupts pending from previous run are not cleared properly,
which results in host receiving an interrupt as soon as it reenables
the NFP but before free list is replenished.

This is caused by a race between reconfiguration of qmon thread and
interrupt generating code.  If we clear pending masks on disable
then the problem cannot occur.