me/msix: add option for just using RX interrupts
For Virtio relay project support for RX queues interrupts is required.
ME MSIX code was not designed for using RX and no TX so if driver
does not configure TX, all that entries will use vector 0 which is
usually reserved for LSC.

Using an extra bit in the reconfig interface could be avoided in case
we need it. It will require more changes in the firmware and likely
also changes to the reset code.