Make declaration of nfd_in_lso_cntr_names in .h file static.
The file me/blocks/vnic/shared/nfd_internal.h in microC is, for some
reason, declaring an array of strings named nfd_in_lso_cntr_names.  Since
the array never gets used in microC this isn't really a problem most of
the time.  But a variable should never be declared in a header file unless
it is either an incomplete declaration or it is declared 'static'.  This
is because multilple source files could include the header file and then
be linked together leading to multiple *separate* instances of the same
variable (one for each object file).  This patch makes the variable
declaration 'static'.