[PCI.OUT] optimize the extration of CTM island in PD
The CTM_ISL field appears in the most significant bits of the descriptor word.
The shift operation can thus simultaneously serve the dual purpose of moving
and masking the required bits.

Use the same type of optimisation on the ctm_and_mu_dma# path, to minimise
differences between the code paths.  This optimisation is not applicable to
the CTM buffer release, as mem[packet_free] takes the island number in the
upper 8 bits.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Signed-off-by: Michael Rapson <michael.rapson@netronome.com>