[PCI.OUT] add support for a skip-free-ctm BLQ
CoreNIC introduces a multicast reference counting scheme based on a special
purpose BLQ. Buffers released to this BLQ can be sent multiple times over,
but NFD must not free the CTM buffer in such cases. A separate ME, the
multicast reaper (MCR), is responsible for deallocating buffers based on
its own internal accounting.

This patch introduces a new compile time option for NFD. NFD_OUT_SKIP_FREE_BLQ
defines the aforementioned BLQ. If the packet's assigned buffer list
matches this BLQ, then the CTM buffer will not be freed by PD. NFD is
still responsible for "freeing" the MU buffer for this packet to the special
BLQ, however, in the CoreNIC use case the buffer is not actually released, but
is rather being sent to MCR via this BLQ for further tracking.

From NFD's perspective, the feature is generic and can be reused by other
projects for their own purposes.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Signed-off-by: Michael Rapson <michael.rapson@netronome.com>