[PCI.IN] save a move by removing q_batch local in notify
The local variable causes nfcc to introduce an unnecessary move
operation in the generated code. Removing the local variable
eliminates this additional cycle.

In the "one at a time" branch, we read out a full batch of descriptors
using the xfers assigned for pkt0.  This means that we must use the
q_num value to compute the qc_queue before we start processing the
batch.

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>
Suggested-by: Edwin Peer <edwin.peer@netronome.com>