[PCI.IN] Pass on set EOP bit to the app
Currently NFD clears word 0, bit 31 of the message to the app MEs.  This
is taking a number of cycles to do.  The descriptor is constructed slightly
more efficiently if we pass it on as received from issue_dma, together
with the offset (meta_len) field.
NFD only passes on messages to the app if EOP bit is set, so passing it on
as a copy with the offset field guarantees it is set in the output.

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>