[PCI.IN] Rename nfd_in_pkt_desc sp0 bit to "is_nfd" and provide test val
The previous patch ensured that sp0 is set to a consistent value,
in an efficient manner for notify.  Now rename that bit and provide a
test value NFD_IN_IS_NFD_TRUE_VAL.  The test value will allow future
optimisation of notify as necessary with a mechanism for apps to check
compatibility between NFD and other components using the PCI.IN output
work queue.
wip://NFDH-305

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>