[cfg] Add nfd_vf_cfg_base function
The new function takes an island and VF number so that it can return the
fully dereferenced address.  It also takes a flag "mb" to select whether
to construct a pointer to the mailbox (on ABI 2) or VF data.  If "mb"
tests true, the VF parameter is ignored.
Adding the function in nfd_vf_cfg_iface.h allows it to be used with
ABI 1 or 2.  When using it with ABI 1, ensure that "mb" is set to false.

Signed-off-by: Johan Moraal <johan.moraal@netronome.com>
Signed-off-by: Michael Rapson <michael.rapson@netronome.com>