[PCI.IN] Move gather related functions from service_qc.c to gather.c, rename
The "vnic setup" and "check queues" functions are more closely associated
with the "gather" component than the "service_qc" function.  During NFD 3.0
implementation, the expectation was that they would be shared by PCI.OUT.
That did not turn out to be the case, so this is some long overdue clean up.

It seems best to move both functions together and to modify the parameters
list of the "check queues" task in a single patch because the state is
tightly coupled.

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>