Create a function from the current tx queue count checking method
The count returned is (still) 16-18 bits so the IRQ mod checks will
trigger prematurely when the TX queue pointer wraps.  This patch
doesn't try to fix that.  It doesn't seem to be bothering anyone yet.
Fixing it while using the QC pointer is not trivial.  The pointer may
wrap before we check the counts so we can't reconstruct a 32 bit
value reliably in the service ME.

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>