[PCI.IN, PCI.OUT] Support NFD_USE_ATMC_CNTRS define to avoid stats engine
We have some platforms without a stats engine, so need a way to avoid
using it.  The NFD_USE_ATMC_CNTRS define in conjunction with recent
flowenv and the NFD_X_CNTRSY_MEM symbols offer developers an option.
This changeset is backward compatible with older flowenv as long as
NFD_USE_ATMC_CNTRS is not used.

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>